[SW-207559] Folder layout refactoring and cleanup (phase 1)
