+++ neural_compressor/torch/algorithms/fp8_quant/__init__.py
@@ -19,7 +19,7 @@ from neural_compressor.torch.algorithms.fp8_quant.common import (
     with_patched_module,
 )
 from neural_compressor.torch.algorithms.fp8_quant.prepare_quant.prepare_model import finish_measurements, prep_model
-from neural_compressor.torch.algorithms.fp8_quant.fp8_quant import FP8Quantizer
+from neural_compressor.torch.algorithms.fp8_quant.quantizer import FP8Quantizer
 from neural_compressor.torch.algorithms.fp8_quant.patched_module_base import (
     PatchedModuleBase,
     register_patched_module,
