+++ neural_compressor/torch/algorithms/fp8_quant/quantizer.py
