+++ neural_compressor/torch/algorithms/fp8_quant/utils/__init__.py
@@ -11,3 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from neural_compressor.torch.algorithms.fp8_quant.utils.patched_module_restore_registry import (
+    helper_mods,
+    helper_mod_register,
+)
