+++ neural_compressor/torch/algorithms/fp8_quant/utils/patched_module_restore_registry.py
